pointless refactoring for readability
